dotbot/firmware: split release fetching out of flash.py into fetch.py
Pure move: flash.py keeps the device-flashing engine and imports fetch_assets/resolve_fw_root back from fetch.py for flash_role's auto-fetch (one-way dep, no cycle). No behavior change.

AI-assisted: Claude Opus 4.8
